First commit of clij.rb itself.
Incomplete, fragmented, unorganized, and showing all the signs
of what it is:  Something I wasn't sure we wanted, wasn't sure
was possible, and the sort of project I have no experience of.

That said, parts of it work, and are even capable of being
useful.  CLI opts, etc will be the immediate goal.