Added methods to enable restarting a job.
With 'Timeout' be untrustworthy we've needed a way for jobs
that tend to choke an overburdended jenkins server to be
restarted with a mininal fuss.  The methods haven't been
implemented everywhere they'll be used ultimately, as testing
some methods is currently impossible.

Also finally took a moment to figure out silly debug default error.